refactor(time): 수행 횟수를 value object로 래핑함
